+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 3.1.0 - Unreleased
+## 3.1.1 - 2026-05-11
 
 ### Security hardening